application/requirements.txt
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
docker==7.1.0
Flask==3.1.2
flask-cors==6.0.2
gunicorn==23.0.0
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.4.0
optional-django==0.3.0
packaging==26.0
PyMySQL==1.1.2
python-dotenv==1.2.1
react==4.3.0
requests==2.32.5
urllib3==2.6.3
Werkzeug==3.1.5
zipp==3.23.0
